Move schemas to separate module. Create core module. Add proto file
