+++ core/src/main/java/eu/alehem/tempserver/remote/core/ThrowingConsumer.java
@@ -1,0 +1,6 @@
+package eu.alehem.tempserver.remote.core;
+
+@FunctionalInterface
+public interface ThrowingConsumer<T, E extends Exception> {
+  void accept(T t) throws E;
+}
