+++ core/src/main/java/eu/alehem/tempserver/remote/core/exceptions/InvalidArgumentsException.java
@@ -1,0 +1,7 @@
+package eu.alehem.tempserver.remote.core.exceptions;
+
+public class InvalidArgumentsException extends Exception {
+  public InvalidArgumentsException(String message) {
+      super(message);
+    }
+}
