+++ core/src/main/java/eu/alehem/tempserver/remote/core/exceptions/ServerCommsFailedException.java
@@ -1,0 +1,7 @@
+package eu.alehem.tempserver.remote.core.exceptions;
+
+public class ServerCommsFailedException extends Exception {
+  public ServerCommsFailedException() {
+    super("ERROR: Server communication failed");
+  }
+}
