+++ core/src/main/java/eu/alehem/tempserver/remote/core/properties/Properties.java
@@ -1,0 +1,5 @@
+package eu.alehem.tempserver.remote.core.properties;
+
+public interface Properties {
+
+}
