+++ core/src/main/java/eu/alehem/tempserver/remote/core/workers/StatusMonitor.java
@@ -1,0 +1,22 @@
+package eu.alehem.tempserver.remote.core.workers;
+
+import eu.alehem.tempserver.remote.core.DatabaseManager;
+import eu.alehem.tempserver.remote.core.TempQueue;
+import java.sql.SQLException;
+import lombok.extern.java.Log;
+
+@Log
+public class StatusMonitor implements Runnable {
+
+  @Override
+  public void run() {
+    TempQueue queue = TempQueue.getInstance();
+    log.info("Queue size: " + queue.getQueueLen());
+    try {
+    log.info("Entries in db: " + DatabaseManager.countMeasurementsInDb());
+    } catch (SQLException s) {
+      log.warning("Failed to count entries in db");
+    }
+  }
+
+}
