+++ pom.xml
@@ -3,19 +3,26 @@
          xmlns="http://maven.apache.org/POM/4.0.0"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
+    <packaging>pom</packaging>
 
     <prerequisites>
         <maven>3.6.0</maven>
     </prerequisites>
+    <modules>
+        <module>schema</module>
+        <module>core</module>
+    </modules>
+
+    <properties>
+        <tempserver.remote.version>1.0.1</tempserver.remote.version>
+        <maven.compiler.release>8</maven.compiler.release>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    </properties>
+
 
     <groupId>eu.alehem.tempserver.remote</groupId>
     <artifactId>tempremote</artifactId>
     <version>1.0.1</version>
-
-    <properties>
-        <maven.compiler.release>8</maven.compiler.release>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    </properties>
 
     <build>
         <pluginManagement>
@@ -50,7 +57,7 @@
                     <configuration>
                         <archive>
                             <manifest>
-                                <mainClass>eu.alehem.tempserver.remote.Main</mainClass>
+                                <mainClass>eu.alehem.tempserver.remote.core.Main</mainClass>
                             </manifest>
                         </archive>
                         <descriptorRefs>
@@ -94,6 +101,13 @@
             <groupId>commons-cli</groupId>
             <artifactId>commons-cli</artifactId>
             <version>1.3.1</version>
+        </dependency>
+
+        <!-- https://mvnrepository.com/artifact/com.google.protobuf/protobuf-java -->
+        <dependency>
+            <groupId>com.google.protobuf</groupId>
+            <artifactId>protobuf-java</artifactId>
+            <version>3.11.4</version>
         </dependency>
 
         <dependency>
