--- src/main/java/eu/alehem/tempserver/remote/ThrowingConsumer.java
@@ -1,6 +1,0 @@
-package eu.alehem.tempserver.remote;
-
-@FunctionalInterface
-public interface ThrowingConsumer<T, E extends Exception> {
-  void accept(T t) throws E;
-}
