--- src/main/java/eu/alehem/tempserver/remote/exceptions/InvalidArgumentsException.java
@@ -1,7 +1,0 @@
-package eu.alehem.tempserver.remote.exceptions;
-
-public class InvalidArgumentsException extends Exception {
-  public InvalidArgumentsException(String message) {
-      super(message);
-    }
-}
