--- src/main/java/eu/alehem/tempserver/remote/exceptions/ServerCommsFailedException.java
@@ -1,7 +1,0 @@
-package eu.alehem.tempserver.remote.exceptions;
-
-public class ServerCommsFailedException extends Exception {
-  public ServerCommsFailedException() {
-    super("ERROR: Server communication failed");
-  }
-}
