--- src/main/java/eu/alehem/tempserver/remote/properties/Properties.java
@@ -1,5 +1,0 @@
-package eu.alehem.tempserver.remote.properties;
-
-public interface Properties {
-
-}
