--- src/main/java/eu/alehem/tempserver/remote/workers/StatusMonitor.java
@@ -1,22 +1,0 @@
-package eu.alehem.tempserver.remote.workers;
-
-import eu.alehem.tempserver.remote.DatabaseManager;
-import eu.alehem.tempserver.remote.TempQueue;
-import java.sql.SQLException;
-import lombok.extern.java.Log;
-
-@Log
-public class StatusMonitor implements Runnable {
-
-  @Override
-  public void run() {
-    TempQueue queue = TempQueue.getInstance();
-    log.info("Queue size: " + queue.getQueueLen());
-    try {
-    log.info("Entries in db: " + DatabaseManager.countMeasurementsInDb());
-    } catch (SQLException s) {
-      log.warning("Failed to count entries in db");
-    }
-  }
-
-}
