--- src/main/resources/config.properties
@@ -1,9 +1,0 @@
-persistence.frequency.seconds=30
-persistence.max_queue_len=60
-persistence.delete_threshold=10000
-persistence.add_threshold=10
-persistence.batch_size=10
-reader.frequency.seconds=120
-sender.frequency.seconds=30
-sender.batch_size=100
-
